chore: Reorganized back-end code into src/ and renamed several files
