+++ backend/package.json
@@ -4,16 +4,17 @@
     "description": "",
     "main": "index.js",
     "scripts": {
+        "strat": "node server.js",
         "test": "jest --watchAll"
     },
     "author": "",
     "license": "ISC",
     "dependencies": {
-        "dotenv": "^5.0.1",
+        "dotenv": "^6.0.0",
         "express": "^4.16.3",
         "node-fetch": "^2.2.0",
         "spotify-web-api-node": "^3.1.1",
-        "ws": "^5.2.2"
+        "ws": "^6.0.0"
     },
     "devDependencies": {
         "jest": "^23.4.2"
