+++ frontend/package.json
@@ -11,26 +11,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/animations": "^6.1.1",
+    "@angular/common": "^6.1.1",
+    "@angular/compiler": "^6.1.1",
+    "@angular/core": "^6.1.1",
+    "@angular/forms": "^6.1.1",
+    "@angular/http": "^6.1.1",
+    "@angular/platform-browser": "^6.1.1",
+    "@angular/platform-browser-dynamic": "^6.1.1",
+    "@angular/router": "^6.1.1",
     "bulma": "^0.7.1",
     "core-js": "^2.5.7",
     "rxjs": "^6.2.2",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^6.1.0",
+    "@angular/compiler-cli": "^6.1.1",
     "@angular-devkit/build-angular": "~0.7.1",
     "typescript": "~2.9.2",
-    "@angular/cli": "~6.1.1",
-    "@angular/language-service": "^6.1.0",
+    "@angular/cli": "~6.1.2",
+    "@angular/language-service": "^6.1.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~10.5.3",
